# product-spec/generated/qgp-api-server/src/Qgp.Api/Controllers/ReportsApi.cs
/*
 * QGP — Quality Governance Portal API
 *
 * Public REST API cho QGP (qgp-api, .NET 8). Sinh từ SDD-QGP-001 §3, map từ PRD-QGP-001 v1.0. Base path /v1. Auth Bearer (OIDC). Verb unsafe cần header Idempotency-Key. Error code theo ISC Internal Standard — API & Error Code. 
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using Qgp.Api.Attributes;
using Qgp.Api.Models;

namespace Qgp.Api.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ReportsApiController : ControllerBase
    { 
        /// <summary>
        /// Compliance acknowledgement (RPT-F-04)
        /// </summary>
        /// <response code="200">Tỉ lệ ack + ai chưa đọc</response>
        /// <response code="400">Lỗi (client/nghiệp vụ) — xem code trong Error</response>
        [HttpGet]
        [Route("/v1/reports/compliance")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("GetComplianceReport")]
        [SwaggerResponse(statusCode: 200, type: typeof(Dictionary<string, Object>), description: "Tỉ lệ ack + ai chưa đọc")]
        [SwaggerResponse(statusCode: 400, type: typeof(Error), description: "Lỗi (client/nghiệp vụ) — xem code trong Error")]
        public virtual IActionResult GetComplianceReport()
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"error\" : {\r\n    \"code\" : \"code\",\r\n    \"message\" : \"message\",\r\n    \"details\" : {\r\n      \"key\" : \"\"\r\n    }\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Dictionary<string, Object>>(exampleJson)
            : new Dictionary<string, Object>();
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Thống kê tài liệu ban hành (RPT-F-01, UC-22)
        /// </summary>
        /// <response code="200">Số liệu aggregate</response>
        /// <response code="400">Lỗi (client/nghiệp vụ) — xem code trong Error</response>
        [HttpGet]
        [Route("/v1/reports/issuance")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("GetIssuanceReport")]
        [SwaggerResponse(statusCode: 200, type: typeof(Dictionary<string, Object>), description: "Số liệu aggregate")]
        [SwaggerResponse(statusCode: 400, type: typeof(Error), description: "Lỗi (client/nghiệp vụ) — xem code trong Error")]
        public virtual IActionResult GetIssuanceReport()
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"error\" : {\r\n    \"code\" : \"code\",\r\n    \"message\" : \"message\",\r\n    \"details\" : {\r\n      \"key\" : \"\"\r\n    }\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Dictionary<string, Object>>(exampleJson)
            : new Dictionary<string, Object>();
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}