# product-spec/generated/qgp-api-server/src/Qgp.Api/wwwroot/openapi-original.json
{
  "openapi" : "3.1.0",
  "info" : {
    "contact" : {
      "name" : "QGP Engineering (SA/TL)"
    },
    "description" : "Public REST API cho QGP (qgp-api, .NET 8). Sinh từ SDD-QGP-001 §3, map từ PRD-QGP-001 v1.0. Base path /v1. Auth Bearer (OIDC). Verb unsafe cần header Idempotency-Key. Error code theo ISC Internal Standard — API & Error Code.\n",
    "license" : {
      "name" : "Proprietary — FPT ISC (internal)",
      "url" : "https://internal.fpt/legal"
    },
    "title" : "QGP — Quality Governance Portal API",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "description" : "Prod (nội bộ)",
    "url" : "https://qgp.internal.fpt/v1"
  }, {
    "description" : "Staging",
    "url" : "https://staging.qgp.internal.fpt/v1"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/reports/issuance" : {
      "get" : {
        "operationId" : "getIssuanceReport",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "additionalProperties" : { }
                }
              }
            },
            "description" : "Số liệu aggregate"
          },
          "400" : {
            "$ref" : "#/components/responses/DefaultError"
          }
        },
        "summary" : "Thống kê tài liệu ban hành (RPT-F-01, UC-22)",
        "tags" : [ "Reports" ]
      }
    },
    "/reports/compliance" : {
      "get" : {
        "operationId" : "getComplianceReport",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "additionalProperties" : { }
                }
              }
            },
            "description" : "Tỉ lệ ack + ai chưa đọc"
          },
          "400" : {
            "$ref" : "#/components/responses/DefaultError"
          }
        },
        "summary" : "Compliance acknowledgement (RPT-F-04)",
        "tags" : [ "Reports" ]
      }
    },
...
  },
  "components" : {
...
    "responses" : {
...
      "DefaultError" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            }
          }
        },
        "description" : "Lỗi (client/nghiệp vụ) — xem code trong Error"
      }
    },
    "schemas" : {
...
      "Error" : {
        "example" : {
          "error" : {
            "code" : "code",
            "message" : "message",
            "details" : {
              "key" : ""
            }
          }
        },
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/Error_error"
          }
        },
        "required" : [ "error" ]
      },
...
      "Error_error" : {
        "example" : {
          "code" : "code",
          "message" : "message",
          "details" : {
            "key" : ""
          }
        },
        "properties" : {
          "code" : {
            "description" : "mã lỗi theo ISC Error Code catalog",
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : true,
            "nullable" : true,
            "type" : "object"
          }
        },
        "required" : [ "code", "message" ]
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "OIDC access token từ SSO nội bộ (ADM-F-05, NFR-04)",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}